Social-To-Lead Agentic Workflow/autostream-agent/requirements.txt
# AutoStream Agent – Python dependencies
# Install with: pip install -r requirements.txt

# LangChain ecosystem
langchain>=0.2.0
langchain-core>=0.2.0
langgraph>=0.1.0

# LLM providers (install at least one)
langchain-openai>=0.1.0           # for GPT-4o-mini
langchain-google-genai>=1.0.0     # for Gemini 1.5 Flash
langchain-anthropic>=0.1.0        # for Claude Haiku

# Utilities
python-dotenv>=1.0.0              # load .env file

# Web server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
